refactor: rename register/Type.tsx to type.ts
