+++ components/auth/Registration/type.ts
